backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
minio==7.1.14
python-dotenv==1.0.0
pydantic==1.10.7

--index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

--index-url https://pypi.org/simple
ultralytics